Move the logic to transform XHTML to NITF into a new class
